extract recref_handler, view_handler
